Implement the build of request_miner request
* Implement a base64Encode(QString) -> QString;

* Implement a base64Encode(QString) -> QString;

* Jsonify the MCTransaction;

* Refactor MCTransaction and MCUser;

* Change transaction and request filenames;

* Disable mining/send moedacoin buttons when the bd is not updated.

+++ include/sendcoindialog.hpp
@@ -17,7 +17,7 @@ public:
 	~SendCoinDialog();
 
 	QString getPubKey();
-	float getValue();
+	double getValue();
 
 private:
 	Ui::SendCoinDialog *ui;
